// src/types/token.ts
export interface Token {
    address: string;
    decimals: number;
    liquidity: number;
    logoURI: string;
    name: string;
    symbol: string;
    volume24hUSD: number;
    volume24hChangePercent: number;
    fdv: number;
    marketcap: number;
    rank: number;
    price: number;
    price24hChangePercent: number;
}